requirements.txt
watchdog
pyyaml
termcolor
cython
pygments
setproctitle
pylru
requests
aioredis
fastapi
uvicorn